# api/openapi/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/v2/events/{eventId}/photos/upload-urls": {
      "post": {
        "operationId": "PhotosController_createUploadUrls",
        "parameters": [
          {
            "name": "eventId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateUploadUrlsDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Upload slots with presigned S3 PUT URLs",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "data",
                    "meta"
                  ],
                  "properties": {
                    "data": {
                      "$ref": "#/components/schemas/UploadSlotResponseDto"
                    },
                    "meta": {
                      "$ref": "#/components/schemas/ResponseMetaDto"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid access token"
          },
          "429": {
            "$ref": "#/components/responses/TooManyRequests"
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "Mint presigned upload URLs for a batch of photos",
        "tags": [
          "photos"
        ]
      }
    },
    "/api/v2/events/{eventId}/photos/confirm": {
      "post": {
        "operationId": "PhotosController_confirmUploads",
        "parameters": [
          {
            "name": "eventId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ConfirmUploadsDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Per-photo verification result",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "data",
                    "meta"
                  ],
                  "properties": {
                    "data": {
                      "$ref": "#/components/schemas/ConfirmPhotoResultDto"
                    },
                    "meta": {
                      "$ref": "#/components/schemas/ResponseMetaDto"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid access token"
          },
          "429": {
            "$ref": "#/components/responses/TooManyRequests"
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "Confirm uploaded photos and mark them ready",
        "tags": [
          "photos"
        ]
      }
    },
    "/api/v2/events/{eventId}/photos": {
      "get": {
        "operationId": "PhotosController_listPhotos",
        "parameters": [
          {
            "name": "eventId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "cursor",
            "required": false,
            "in": "query",
            "description": "Opaque cursor: the nextCursor value from the previous page. Omit for the first page.",
            "schema": {
              "maxLength": 255,
              "type": "string"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 1,
              "maximum": 100,
              "default": 50,
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Photos with presigned download URLs, newest first",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "data",
                    "meta"
                  ],
                  "properties": {
                    "data": {
                      "$ref": "#/components/schemas/PhotoListResponseDto"
                    },
                    "meta": {
                      "$ref": "#/components/schemas/ResponseMetaDto"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid access token"
          },
          "429": {
            "$ref": "#/components/responses/TooManyRequests"
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "List ready photos in an event (cursor-paginated)",
        "tags": [
          "photos"
        ]
      }
    },
    "/api/v2/photos/{photoId}": {
      "get": {
        "operationId": "PhotosController_findOne",
        "parameters": [
          {
            "name": "photoId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Photo with presigned download URL",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "data",
                    "meta"
                  ],
                  "properties": {
                    "data": {
                      "$ref": "#/components/schemas/PhotoResponseDto"
                    },
                    "meta": {
                      "$ref": "#/components/schemas/ResponseMetaDto"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid access token"
          },
          "429": {
            "$ref": "#/components/responses/TooManyRequests"
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "Get a photo by ID",
        "tags": [
          "photos"
        ]
      },
      "delete": {
        "operationId": "PhotosController_remove",
        "parameters": [
          {
            "name": "photoId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Photo deleted (empty data envelope at runtime)"
          },
          "401": {
            "description": "Missing or invalid access token"
          },
          "429": {
            "$ref": "#/components/responses/TooManyRequests"
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "Delete a photo",
        "tags": [
          "photos"
        ]
      }
    },
...
  },
  "info": {
    "title": "Everglow API",
    "description": "Photo-sharing platform for events — HTTP API contract (OpenAPI 3)",
    "version": "2.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "access-token": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http",
        "description": "JWT access token"
      }
    },
    "schemas": {
...
      "ResponseMetaDto": {
        "type": "object",
        "properties": {
          "timestamp": {
            "type": "string",
            "example": "2026-06-03T12:00:00.000Z"
          },
          "path": {
            "type": "string",
            "example": "/api/v2/auth/signin"
          }
        },
        "required": [
          "timestamp",
          "path"
        ]
      },
...
      "UploadSlotResponseDto": {
        "type": "object",
        "properties": {
          "photoId": {
            "type": "string",
            "format": "uuid"
          },
          "uploadUrl": {
            "type": "string",
            "maxLength": 2048,
            "description": "Presigned S3 PUT URL the client uploads bytes to"
          }
        },
        "required": [
          "photoId",
          "uploadUrl"
        ]
      },
      "UploadFileDto": {
        "type": "object",
        "properties": {
          "contentType": {
            "type": "string",
            "enum": [
              "image/jpeg",
              "image/png",
              "image/webp",
              "image/heic",
              "image/heif"
            ],
            "example": "image/jpeg"
          },
          "sizeBytes": {
            "type": "number",
            "minimum": 1,
            "maximum": 26214400,
            "example": 1048576
          }
        },
        "required": [
          "contentType",
          "sizeBytes"
        ]
      },
      "CreateUploadUrlsDto": {
        "type": "object",
        "properties": {
          "files": {
            "maxItems": 20,
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/UploadFileDto"
            }
          }
        },
        "required": [
          "files"
        ]
      },
      "ConfirmPhotoResultDto": {
        "type": "object",
        "properties": {
          "photoId": {
            "type": "string",
            "format": "uuid"
          },
          "status": {
            "type": "string",
            "enum": [
              "READY",
              "MISSING",
              "MISMATCHED",
              "NOT_FOUND"
            ],
            "example": "READY",
            "description": "READY: the object was verified and the photo is now visible. MISSING: no object was uploaded; the slot has been released, mint a new one. MISMATCHED: the object differs from the declared size or type; it and the slot have been removed, mint a new one. NOT_FOUND: not a pending upload of the caller in this event."
          }
        },
        "required": [
          "photoId",
          "status"
        ]
      },
      "ConfirmUploadsDto": {
        "type": "object",
        "properties": {
          "photoIds": {
            "maxItems": 20,
            "type": "array",
            "items": {
              "type": "string",
              "format": "uuid"
            }
          }
        },
        "required": [
          "photoIds"
        ]
      },
      "PhotoResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "eventId": {
            "type": "string",
            "format": "uuid"
          },
          "addedById": {
            "type": "string",
            "format": "uuid",
            "nullable": true,
            "description": "Who uploaded the photo; null once that account has been deleted and the photo was kept for the event."
          },
          "url": {
            "type": "string",
            "maxLength": 2048,
            "description": "Presigned S3 GET URL, valid for a short period"
          },
          "contentType": {
            "type": "string",
            "enum": [
              "image/jpeg",
              "image/png",
              "image/webp",
              "image/heic",
              "image/heif"
            ],
            "example": "image/jpeg"
          },
          "createdAt": {
            "format": "date-time",
            "type": "string"
          }
        },
        "required": [
          "id",
          "eventId",
          "addedById",
          "url",
          "contentType",
          "createdAt"
        ]
      },
      "PhotoListResponseDto": {
        "type": "object",
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/PhotoResponseDto"
            }
          },
          "nextCursor": {
            "type": "string",
            "nullable": true,
            "description": "Opaque cursor for the next page; pass it as ?cursor=. Null on the last page."
          }
        },
        "required": [
          "items",
          "nextCursor"
        ]
      },
...
    },
    "responses": {
      "TooManyRequests": {
        "description": "Rate limit exceeded; retry after the number of seconds in the Retry-After header",
        "headers": {
          "Retry-After": {
            "description": "Seconds to wait before retrying",
            "schema": {
              "type": "integer",
              "minimum": 1
            }
          }
        },
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "required": [
                "meta"
              ],
              "properties": {
                "message": {
                  "type": "string",
                  "example": "Too many requests, please try again later"
                },
                "code": {
                  "type": "string",
                  "description": "Stable machine-readable error code, when the error has one",
                  "example": "RATE_LIMIT_EXCEEDED"
                },
                "meta": {
                  "$ref": "#/components/schemas/ResponseMetaDto"
                }
              }
            }
          }
        }
      }
    }
  }
}